Update MicroBooNE.tex
Completed commenting on it.  Two citations are missing. I just put them in to ensure the flagging.  I have a problem with the claim in the figure caption since I do not quite see what is mentioned in it, such as missing SciBooNE results in the plot and no T2K data points at high energies, etc.